Переместит кнопки действий из поста в отдельный компонент
